Client/requirements.txt
paho-mqtt==1.6.1
Pillow==11.2.1
pyobjc-framework-Quartz==9.2 ; sys_platform == 'darwin' 